Get the starter code to work with the Udacity Workspace
