+++ README.md
@@ -29,13 +29,18 @@ Remember that once you clone, you will still need to install everything:
 
 ## Setting up the API
 
-The Aylien API is perhaps different than what you've used before. It has you install a node module to run certain commands through, it will simplify the requests we need to make from our node/express backend.
+The Aylien/MeaningCloud API is perhaps different than what you've used before. You may need to install a node module to run certain commands through, it will simplify the requests we need to make from our node/express backend.
 
 ### Step 1: Signup for an API key
-First, you will need to go [here](https://developer.aylien.com/signup). Signing up will get you an API key. Don't worry, at the time of this course, the API is free to use up to 1000 requests per day or 333 intensive requests. It is free to check how many requests you have remaining for the day.
 
-### Step 2: Install the SDK
-Next you'll need to get the SDK. SDK stands for Software Development Kit, and SDKs are usually a program that brings together various tools to help you work with a specific technology. SDKs will be available for all the major languages and platforms, for instance the Aylien SDK brings together a bunch of tools and functions that will make it possible to interface with their API from our server and is available for Node, Python, PHP, Go, Ruby and many others. We are going to use the Node one, the page is available [here](https://docs.aylien.com/textapi/sdks/#sdks). You get 1000 free requests per day. 
+**For the Aylien API:** First, you will need to go [here](https://developer.aylien.com/signup). Signing up will get you an API key. Don't worry, at the time of this course, the API is free to use up to 1000 requests per day or 333 intensive requests. It is free to check how many requests you have remaining for the day.
+
+**For the MeaningCloud API**: You can find the API [here](https://www.meaningcloud.com/developer/sentiment-analysis). Once you create an account with MeaningCloud, you will be given a license key to start using the API. This API does not require an SDK, so you can skip ahead to step 4 in the instructions.
+
+### Step 2: Install the SDK (Aylien API only)
+Next you'll need to get the SDK. SDK stands for Software Development Kit, and SDKs are usually a program that brings together various tools to help you work with a specific technology. SDKs will be available for all the major languages and platforms, for instance the Aylien SDK brings together a bunch of tools and functions that will make it possible to interface with their API from our server and is available for Node, Python, PHP, Go, Ruby and many others. We are going to use the Node one, the page is available [here](https://docs.aylien.com/textapi/sdks/#sdks). You get 1000 free requests per day.
+
+**For the MeaningCloud API**, you may need to install the [form-data](https://www.npmjs.com/package/form-data) module to follow its [JavaScript example](https://learn.meaningcloud.com/developer/sentiment-analysis/2.1/dev-tools).
 
 ### Step 3: Require the SDK package
 Install the SDK in your project and then we'll be ready to set up your server/index.js file.
@@ -99,9 +104,9 @@ I won't provide further examples here, as it's up to you to create the various r
 Once you are hooked up to the Aylien API, you are most of the way there! Along with making sure you are following all the requirements in the project rubric in the classroom, here are a few other steps to make sure you take.
 
 - Parse the response body to dynamically fill content on the page.
-- Test that the server and form submission work, making sure to also handle error responses if the user input does not match API requirements. 
-- Go back to the web pack config and add the setup for service workers.  
-- Test that the site is now available even when you stop your local server 
+- Test that the server and form submission work, making sure to also handle error responses if the user input does not match API requirements.
+- Go back to the web pack config and add the setup for service workers. 
+- Test that the site is now available even when you stop your local server
 
 ## Deploying
