--- starter project/src/client/js/formHandler.js
@@ -1,23 +1,0 @@
-//Make fnction that check url and import it here
-
-const form = document.getElementById('urlForm');
-form.addEventListener('submit', handleSubmit);
-
-function handleSubmit(event) {
-    event.preventDefault();
-
-    // Get the URL from the input field
-    const formText = document.getElementById('name').value;
-    
-    // Check if the URL is valid
- 
-        // If the URL is valid, send it to the server
-      
-    
-}
-
-// Function to send data to the server
-
-
-export { handleSubmit };
-
