+++ starter_project/package.json
@@ -6,13 +6,14 @@
     "test": "jest",
     "start": "node src/server/index.js",
     "build-prod": "webpack --config webpack.prod.js",
-    "build-dev": "webpack-dev-server  --config webpack.dev.js --open"
+    "build-dev": "webpack-dev-server  --config webpack.dev.js"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "description": "",
   "dependencies": {
+    "clean-webpack-plugin": "^4.0.0",
     "cors": "^2.8.5",
     "css-loader": "^6.10.0",
     "dotenv": "^16.4.5",
